requirements.txt
aiohappyeyeballs
aiohttp
aioresponses
aiosignal
attrs
beautifulsoup4
certifi
charset-normalizer
dnspython
frozenlist
httpretty
idna
lxml
multidict
packaging
propcache
py-mini-racer
python-dateutil
python-Wappalyzer
python-whois
requests
setuptools
six
soupsieve
typing_extensions
urllib3
yarl
